+++ example/lib/main.dart
@@ -1,4 +1,5 @@
 //
+
 import 'package:fluttery_framework_example/src/view.dart';
 
 void main() => runApp(TemplateApp());